// data/test-data/href-values.enum.ts
export enum Hrefs {
    BANK_BRANCHES = "/map",
    EXCHANGE_RATES = "/exchange-rates",
    CONTACT = "/contact",
    PRIVACY_POLICY = "/documents/privacy-policy",
    REMOTE_BANKING_RULES = "/documents/remote-banking-rules",
    COOKIES_POLICY = "/documents/personal-data-terms",
    TEL_NUMBER = "tel:+44 (0)20 3461 4444",
    EMAIL = "mailto:contact@polybank.co.uk",
    TERMS_AND_CONDITIONS = "/documents",
    GOOGLE_PLAY = "https://play.google.com/",
    APPLE_STORE = "https://apps.apple.com/",
    RECOVERY_PASSWORD = "/recovery-password",
    SIGNUP = "/signup",
    MAIN_MENU = "/main-menu",
}