// data/text-data/contact.enum.ts
export enum ContactTextData {
    TITLE = "Contact us",
    FOR_INDIVIDUALS_TITLE = "For individuals",
    PRODUCT_ADVICE = "Product advice, account and transaction information",
    EVERY_DAY = "Every day from 8:00 to 20:00",
    CONTACTS = "Contacts",
    CARDS_SUPPORT = "Cards support",
    NEW_CARD = "A new card, credit card charges, report lost or stolen",
    WEEK_DAYS = "Weekdays from 8:00 to 20:00",
    CONTACTS_CARDS = "Contacts",
}